farmconnect/backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Pillow==9.2.0
django-cors-headers==3.13.0
stripe==5.0.0
celery==5.2.7
redis==4.3.4